1.41

+++ project/src/main/java/com/mic/tech/DeleteFilmAction.java
@@ -5,9 +5,9 @@ import java.util.Map;
 import java.util.Scanner;
 
 public class DeleteFilmAction extends AbstractAuthenticatedAction{
-    FilmService filmService=null;
-    GlobalState state=null;
-    Scanner scanner=null;
+    private FilmService filmService=null;
+    private GlobalState state=null;
+    private Scanner scanner=null;
     DeleteFilmAction(GlobalState state,FilmService filmService,Scanner scanner){
         this.state=state;
         this.filmService=filmService;
@@ -27,7 +27,10 @@ public class DeleteFilmAction extends AbstractAuthenticatedAction{
             String title = scanner.nextLine();
             Film film = filmService.getFilmByFilmTitle(title);
             if (film != null) {
-                filmService.deleteFilm(title);
+                super.println("删除后无法恢复，请用户确认是否继续删除操作。y/n");
+                String s=scanner.nextLine();
+                if(s.equals("y"))
+                    filmService.deleteFilm(title);
             }
             else {
                 super.println("找不到该影片");
